[streaming] new tests in MapTest

+++ flink-addons/flink-streaming/src/main/java/eu/stratosphere/streaming/api/streamcomponent/StreamWindowTask.java
@@ -53,7 +53,7 @@ public class StreamWindowTask extends FlatMapFunction<Tuple, Tuple> {
 
 	@Override
 	public void flatMap(Tuple value, Collector<Tuple> out) throws Exception {
-		long progress = (Long) value.getField(windowFieldId);
+		long progress = value.getField(windowFieldId);
 		if (initTimestamp == -1) {
 			initTimestamp = progress;
 			nextTimestamp = initTimestamp + computeGranularity;
